# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CalendarEvent {
  id: string;
  title: string;
  start: string;
  end: string;
  allDay: boolean;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  description?: string | null;
  calendarId?: string;
  /** @nullable */
  calendarName?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  calendarColor?: string | null;
}

export type CalendarMetaProvider = typeof CalendarMetaProvider[keyof typeof CalendarMetaProvider];


export const CalendarMetaProvider = {
  google: 'google',
  outlook: 'outlook',
} as const;

export interface CalendarMeta {
  id: string;
  name: string;
  color: string;
  accessRole: string;
  selected: boolean;
  provider: CalendarMetaProvider;
}

export interface CalendarPref {
  visible: boolean;
  /** @nullable */
  color?: string | null;
}

export interface CalendarAccount {
  id: number;
  email: string;
}

export interface CalendarStatus {
  connected: boolean;
  provider: string;
  outlookConnected?: boolean;
  googleAccounts?: CalendarAccount[];
}

export interface ErrorResponse {
  error: string;
}

export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority];


export const TaskPriority = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface Task {
  id: string;
  title: string;
  project: string;
  due: string;
  time?: string;
  estTime?: string;
  priority: TaskPriority;
  done: boolean;
}

export interface Project {
  id: string;
  name: string;
  description: string;
  color: string;
  goal: string;
}

export interface Capture {
  id: string;
  text: string;
  createdAt: string;
  converted: boolean;
}

export type PreferencesFontStyle = typeof PreferencesFontStyle[keyof typeof PreferencesFontStyle];


export const PreferencesFontStyle = {
  modern: 'modern',
  classic: 'classic',
  rounded: 'rounded',
} as const;

export type PreferencesCalendarConnected = typeof PreferencesCalendarConnected[keyof typeof PreferencesCalendarConnected];


export const PreferencesCalendarConnected = {
  none: 'none',
  google: 'google',
  outlook: 'outlook',
} as const;

export type PreferencesCalendarPrefs = {[key: string]: CalendarPref};

export interface Preferences {
  dark: boolean;
  accent: string;
  memory: boolean;
  reminders: boolean;
  sectionOrder: string[];
  fontStyle: PreferencesFontStyle;
  calendarConnected: PreferencesCalendarConnected;
  calendarPrefs?: PreferencesCalendarPrefs;
  dismissedDuplicates?: string[];
}

export interface WaitingFor {
  id: string;
  person: string;
  item: string;
  addedAt: string;
}

export interface ImportantDate {
  id: string;
  label: string;
  date: string;
}

export interface ConnectionLog {
  id: string;
  date: string;
  note?: string;
  method?: string;
}

export type PersonFrequency = typeof PersonFrequency[keyof typeof PersonFrequency];


export const PersonFrequency = {
  weekly: 'weekly',
  biweekly: 'biweekly',
  monthly: 'monthly',
  custom: 'custom',
  none: 'none',
} as const;

export interface Person {
  id: string;
  name: string;
  relationship: string;
  contactMethod: string;
  frequency: PersonFrequency;
  customDays?: number;
  birthday?: string;
  importantDates: ImportantDate[];
  notes: string;
  lastConnectedAt?: string;
  reminderSnoozedUntil?: string;
  connections: ConnectionLog[];
}

export interface AppState {
  tasks: Task[];
  projects: Project[];
  captures: Capture[];
  preferences: Preferences;
  waitingFor: WaitingFor[];
  people: Person[];
}

export type GetCalendarEventsParams = {
/**
 * ISO date string (YYYY-MM-DD), defaults to today
 */
date?: string;
};

export type GetCalendarWeekSummaryParams = {
/**
 * ISO date string (YYYY-MM-DD) for the Monday of the week
 */
weekStart: string;
};

export type GetCalendarWeekSummary200 = {[key: string]: number};