# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: calendar
    description: Calendar operations
  - name: app
    description: App state operations
  - name: auth
    description: Authentication and connected account management
paths:
  /auth/accounts:
    get:
      operationId: getAuthAccounts
      tags: [auth]
      summary: List connected Google accounts for this session
      responses:
        "200":
          description: List of connected accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarAccount"
  /auth/accounts/{id}:
    delete:
      operationId: deleteAuthAccountsId
      tags: [auth]
      summary: Disconnect a Google account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Account disconnected
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /calendar/events:
    get:
      operationId: getCalendarEvents
      tags: [calendar]
      summary: Get calendar events for a date range
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
          description: ISO date string (YYYY-MM-DD), defaults to today
      responses:
        "200":
          description: List of calendar events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEvent"
        "503":
          description: Calendar not connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /calendar/status:
    get:
      operationId: getCalendarStatus
      tags: [calendar]
      summary: Check if calendar is connected
      responses:
        "200":
          description: Calendar connection status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarStatus"
  /calendar/week-summary:
    get:
      operationId: getCalendarWeekSummary
      tags: [calendar]
      summary: Get event counts per day for a week
      parameters:
        - name: weekStart
          in: query
          required: true
          schema:
            type: string
          description: ISO date string (YYYY-MM-DD) for the Monday of the week
      responses:
        "200":
          description: Map of date strings to event counts
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: number
        "503":
          description: Calendar not connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /calendar/calendars:
    get:
      operationId: getCalendarList
      tags: [calendar]
      summary: List all calendars the user has access to
      responses:
        "200":
          description: List of calendars
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarMeta"
        "503":
          description: Calendar not connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /app/state:
    get:
      operationId: getAppState
      tags: [app]
      summary: Get persisted app state
      description: Returns the saved app state. If no state has been saved yet, returns the seed state and persists it.
      responses:
        "200":
          description: App state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppState"
    put:
      operationId: saveAppState
      tags: [app]
      summary: Save app state
      description: Replaces the entire persisted app state.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppState"
      responses:
        "200":
          description: Saved app state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppState"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    CalendarEvent:
      type: object
      required: [id, title, start, end, allDay]
      properties:
        id:
          type: string
        title:
          type: string
        start:
          type: string
        end:
          type: string
        allDay:
          type: boolean
        location:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        calendarId:
          type: string
        calendarName:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        calendarColor:
          type: ["string", "null"]
    CalendarMeta:
      type: object
      required: [id, name, color, accessRole, selected, provider]
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
        accessRole:
          type: string
        selected:
          type: boolean
        provider:
          type: string
          enum: [google, outlook]
    CalendarPref:
      type: object
      required: [visible]
      properties:
        visible:
          type: boolean
        color:
          type: ["string", "null"]
    CalendarAccount:
      type: object
      required: [id, email]
      properties:
        id:
          type: number
        email:
          type: string
    CalendarStatus:
      type: object
      required: [connected, provider]
      properties:
        connected:
          type: boolean
        provider:
          type: string
        outlookConnected:
          type: boolean
        googleAccounts:
          type: array
          items:
            $ref: "#/components/schemas/CalendarAccount"
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Task:
      type: object
      required: [id, title, project, due, priority, done]
      properties:
        id:
          type: string
        title:
          type: string
        project:
          type: string
        due:
          type: string
        time:
          type: string
        estTime:
          type: string
        priority:
          type: string
          enum: [high, medium, low]
        done:
          type: boolean
    Project:
      type: object
      required: [id, name, description, color, goal]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        color:
          type: string
        goal:
          type: string
    Capture:
      type: object
      required: [id, text, createdAt, converted]
      properties:
        id:
          type: string
        text:
          type: string
        createdAt:
          type: string
        converted:
          type: boolean
    Preferences:
      type: object
      required: [dark, accent, memory, reminders, sectionOrder, fontStyle, calendarConnected]
      properties:
        dark:
          type: boolean
        accent:
          type: string
        memory:
          type: boolean
        reminders:
          type: boolean
        sectionOrder:
          type: array
          items:
            type: string
        fontStyle:
          type: string
          enum: [modern, classic, rounded]
        calendarConnected:
          type: string
          enum: [none, google, outlook]
        calendarPrefs:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/CalendarPref"
        dismissedDuplicates:
          type: array
          items:
            type: string
    WaitingFor:
      type: object
      required: [id, person, item, addedAt]
      properties:
        id:
          type: string
        person:
          type: string
        item:
          type: string
        addedAt:
          type: string
    ImportantDate:
      type: object
      required: [id, label, date]
      properties:
        id:
          type: string
        label:
          type: string
        date:
          type: string
    ConnectionLog:
      type: object
      required: [id, date]
      properties:
        id:
          type: string
        date:
          type: string
        note:
          type: string
        method:
          type: string
    Person:
      type: object
      required: [id, name, relationship, contactMethod, frequency, importantDates, notes, connections]
      properties:
        id:
          type: string
        name:
          type: string
        relationship:
          type: string
        contactMethod:
          type: string
        frequency:
          type: string
          enum: [weekly, biweekly, monthly, custom, none]
        customDays:
          type: number
        birthday:
          type: string
        importantDates:
          type: array
          items:
            $ref: "#/components/schemas/ImportantDate"
        notes:
          type: string
        lastConnectedAt:
          type: string
        reminderSnoozedUntil:
          type: string
        connections:
          type: array
          items:
            $ref: "#/components/schemas/ConnectionLog"
    AppState:
      type: object
      required: [tasks, projects, captures, preferences, waitingFor, people]
      properties:
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        captures:
          type: array
          items:
            $ref: "#/components/schemas/Capture"
        preferences:
          $ref: "#/components/schemas/Preferences"
        waitingFor:
          type: array
          items:
            $ref: "#/components/schemas/WaitingFor"
        people:
          type: array
          items:
            $ref: "#/components/schemas/Person"