# lib/api-zod/src/generated/types/calendarStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CalendarAccount } from './calendarAccount';

export interface CalendarStatus {
  connected: boolean;
  provider: string;
  outlookConnected?: boolean;
  googleAccounts?: CalendarAccount[];
}